fix(chain): address authority index review
