+++ packages/chain/src/index.ts
@@ -6,7 +6,7 @@ export {
 } from './context-graph-authority-history.js';
 export {
   type ContextGraphAuthorityIndexStore,
-} from './context-graph-authority-index.js';
+} from './context-graph-authority-index-checkpoint.js';
 export {
   bindContextGraphAuthorityReader,
   type ContextGraphAuthorityReader,
